VR: separate structure elements
